[common] - extract go generation logic to shared package - extract common model-annotator - support for omitempty in the shared generator and model-annotator logic - added required indentation style for Makefile in the .editorconfig
[tekton-annotator]
- replace tekton-annotator with shared model-annotator

[tekton-generator]
- generate separate tekton schema for versions using different go source dependency version
- use shared go generation logic for tekton generators
- use go vendor directory for tekton generators
- skip building the generator binary for tekton, instead only run the code

[tekton-model]
- split tekton model by version into v1alpha1 and v1beta1
- remove the generated *Schema class for tekton
