+++ extensions/tekton/client/pom.xml
@@ -54,7 +54,11 @@
     </dependency>
     <dependency>
       <groupId>io.fabric8</groupId>
-      <artifactId>tekton-model</artifactId>
+      <artifactId>tekton-model-v1alpha1</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>io.fabric8</groupId>
+      <artifactId>tekton-model-v1beta1</artifactId>
     </dependency>
     <dependency>
       <groupId>io.fabric8</groupId>
