+++ extensions/tekton/generator-v1alpha1/.gitignore
@@ -1,0 +1,1 @@
+/vendor/
