+++ extensions/tekton/generator-v1beta1/.gitignore
@@ -1,0 +1,1 @@
+/vendor/
