--- extensions/tekton/generator/.gitignore
@@ -1,1 +1,0 @@
-generate
