--- extensions/tekton/generator/go.mod
@@ -1,15 +1,0 @@
-module github.com/fabric8io/kubernetes-client/extensions/tekton/generator
-
-require (
-	github.com/gregjones/httpcache v0.0.0-20190611155906-901d90724c79 // indirect
-	github.com/knative/build v0.7.0 // indirect
-	github.com/knative/eventing v0.7.1
-	github.com/knative/pkg v0.0.0-20190617142447-13b093adc272 // indirect
-	github.com/knative/serving v0.6.0
-	github.com/knative/test-infra v0.0.0-20190702025031-91d37e4abc30 // indirect
-	github.com/tektoncd/pipeline v0.11.0
-	k8s.io/apimachinery v0.17.2
-	k8s.io/cli-runtime v0.0.0-20190325194458-f2b4781c3ae1 // indirect
-)
-
-go 1.13
