+++ extensions/tekton/model-v1alpha1/src/main/java/io/fabric8/tekton/pipeline/v1alpha1/ArrayOrString.java
