--- extensions/tekton/model/pom.xml
@@ -1,153 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2015 Red Hat, Inc.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-            http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-      <groupId>io.fabric8</groupId>
-      <artifactId>tekton-extension-pom</artifactId>
-      <version>4.9-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>tekton-model</artifactId>
-    <packaging>bundle</packaging>
-    <name>Fabric8 :: Tekton :: Model</name>
-
-    <properties>
-        <osgi.import>
-            io.fabric8.kubernetes.api.builder,
-            !io.fabric8.tekton.*,
-            *
-        </osgi.import>
-        <osgi.export>
-            io.fabric8.tekton,
-            io.fabric8.tekton.*
-        </osgi.export>
-        <osgi.include.resources>
-          {maven-resources},
-          /tekton.properties=target/classes/tekton.properties
-        </osgi.include.resources>
-    </properties>
-
-    <dependencies>
-      <dependency>
-        <groupId>io.sundr</groupId>
-        <artifactId>builder-annotations</artifactId>
-      </dependency>
-      <dependency>
-        <groupId>io.sundr</groupId>
-        <artifactId>transform-annotations</artifactId>
-      </dependency>
-      <dependency>
-        <groupId>io.fabric8</groupId>
-        <artifactId>kubernetes-model-core</artifactId>
-      </dependency>
-      <dependency>
-        <groupId>io.fabric8</groupId>
-        <artifactId>knative-model</artifactId>
-      </dependency>
-      <dependency>
-        <groupId>com.fasterxml.jackson.module</groupId>
-        <artifactId>jackson-module-jaxb-annotations</artifactId>
-      </dependency>
-      <dependency>
-        <groupId>org.junit.jupiter</groupId>
-        <artifactId>junit-jupiter-engine</artifactId>
-        <scope>test</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.junit.jupiter</groupId>
-        <artifactId>junit-jupiter-migrationsupport</artifactId>
-        <scope>test</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.skyscreamer</groupId>
-        <artifactId>jsonassert</artifactId>
-        <scope>test</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.projectlombok</groupId>
-        <artifactId>lombok</artifactId>
-        <scope>provided</scope>
-      </dependency>
-    </dependencies>
-
-    <build>
-      <plugins>
-        <plugin>
-          <groupId>org.jsonschema2pojo</groupId>
-          <artifactId>jsonschema2pojo-maven-plugin</artifactId>
-          <version>${jsonschema2pojo.version}</version>
-          <configuration>
-            <sourceDirectory>${project.basedir}/src/main/resources/schema</sourceDirectory>
-            <targetPackage>io.fabric8.tekton.api.model</targetPackage>
-            <includeConstructors>true</includeConstructors>
-            <includeJsr303Annotations>false</includeJsr303Annotations>
-            <includeToString>false</includeToString>
-            <includeHashcodeAndEquals>false</includeHashcodeAndEquals>
-            <outputDirectory>${project.build.directory}/generated-sources</outputDirectory>
-            <customAnnotator>io.fabric8.tekton.TektonTypeAnnotator</customAnnotator>
-          </configuration>
-          <executions>
-            <execution>
-              <id>generate</id>
-              <phase>generate-sources</phase>
-              <goals>
-                <goal>generate</goal>
-              </goals>
-            </execution>
-          </executions>
-          <dependencies>
-            <dependency>
-              <groupId>io.fabric8</groupId>
-              <artifactId>tekton-model-annotator</artifactId>
-              <version>${project.version}</version>
-            </dependency>
-          </dependencies>
-        </plugin>
-        <plugin>
-          <groupId>org.codehaus.mojo</groupId>
-          <artifactId>build-helper-maven-plugin</artifactId>
-          <version>${maven.buildhelper.plugin.version}</version>
-          <executions>
-            <execution>
-              <id>attach-artifacts</id>
-              <phase>package</phase>
-              <goals>
-                <goal>attach-artifact</goal>
-              </goals>
-              <configuration>
-                <artifacts>
-                  <artifact>
-                    <file>${project.build.outputDirectory}/schema/tekton-schema.json</file>
-                    <type>json</type>
-                    <classifier>schema</classifier>
-                  </artifact>
-                </artifacts>
-              </configuration>
-            </execution>
-          </executions>
-        </plugin>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-release-plugin</artifactId>
-        </plugin>
-      </plugins>
-    </build>
-
-</project>
