+++ generator/go.mod
@@ -1,0 +1,3 @@
+module github.com/fabric8io/kubernetes-client/generator
+
+go 1.13
