+++ generator/pkg/schemagen/json.go
